Fix the (wrong) mix of parameter conventions for the Gamma distribution
jStat documentation:
> Returns the value of `x` in the pdf of the Gamma distribution with the
> parameters shape (`k`) and scale (`theta`). Notice that if using the
> alpha beta convention, `scale = 1/beta`.

The formula, and calculation of the mean and variance however assumed
the convention of β being the rate, and β was plugged into jStat
functions despite being a rate (and not the scale parameter).

Fix this by renaming β to λ (which makes Gamma(1, λ) the Exponential
distribution and makes clear which parameter is which), and correctly
plug in the `scale` parameters to jStat.