Fix a Distributions update bug
When entering the URL, the distribution object would be correctly
initialized from parameters, but these would then not be stored in the
URL again, leading to the problem that reloading once would reset the
parameters to default, and reloading twice switched to the Normal(0,1)
distribution.

The fix is simply updating the URL after everything has been properly
initialized.